requirements.txt
requests
absl-py
numpy
tensorflow-gpu
opencv-python
tqdm
lxml
matplotlib
certifi
cycler
kiwisolver
matplotlib
pyparsing
PyQt5
PyQt5-sip
PyQtWebEngine
python-dateutil
six
tornado
